+++ spec/ReportTransfer.spec.php
@@ -12,9 +12,9 @@
 namespace coverallskit\spec;
 
 use coverallskit\ReportTransfer;
-use coverallskit\entity\Report;
-use coverallskit\entity\Service;
-use coverallskit\entity\Repository;
+use coverallskit\entity\CoverallsReport;
+use coverallskit\entity\CIService;
+use coverallskit\entity\GitRepository;
 use coverallskit\entity\collection\SourceFileCollection;
 use coverallskit\Environment;
 use coverallskit\environment\TravisCI;
@@ -44,12 +44,12 @@ describe('ReportTransfer', function() {
                 'COVERALLS_REPO_TOKEN' => 'token'
             ]);
             $adaptor = new TravisCI($environment);
-            $service = new Service($adaptor);
+            $service = new CIService($adaptor);
 
-            $this->report = new Report([
+            $this->report = new CoverallsReport([
                 'name' => __DIR__ . '/fixtures/coveralls.json',
                 'token' => 'foo',
-                'repository' => new Repository(__DIR__ . '/../'),
+                'repository' => new GitRepository(__DIR__ . '/../'),
                 'service' => $service,
                 'sourceFiles' => new SourceFileCollection()
             ]);