+++ spec/entity/CoverallsReport.spec.php
@@ -11,14 +11,14 @@
 
 namespace coverallskit\spec;
 
-use coverallskit\entity\Report;
-use coverallskit\entity\Repository;
+use coverallskit\entity\CoverallsReport;
+use coverallskit\entity\GitRepository;
 use coverallskit\entity\SourceFile;
 use coverallskit\entity\collection\SourceFileCollection;
 use Prophecy\Prophet;
 
 
-describe('Report', function() {
+describe('CoverallsReport', function() {
     beforeEach(function() {
         $this->path = __DIR__ . '/tmp/coverage.json';
 
@@ -31,28 +31,28 @@ describe('Report', function() {
     describe('isEmpty', function() {
         context('when empty', function() {
             it('should return true', function () {
-                $report = new Report();
+                $report = new CoverallsReport();
                 expect($report->isEmpty())->toBeTrue();
             });
         });
     });
     describe('token', function() {
         it('should return repository token string', function() {
-            $report = new Report([ 'token' => 'foo' ]);
+            $report = new CoverallsReport([ 'token' => 'foo' ]);
             expect($report->getToken())->toBe('foo');
         });
     });
     describe('repository', function() {
         it('should return repository', function() {
-            $report = new Report([
-                'repository' => new Repository(__DIR__ . '/../../')
+            $report = new CoverallsReport([
+                'repository' => new GitRepository(__DIR__ . '/../../')
             ]);
-            expect($report->getRepository())->toBeAnInstanceOf('coverallskit\entity\Repository');
+            expect($report->getRepository())->toBeAnInstanceOf('coverallskit\entity\RepositoryEntity');
         });
     });
     describe('sourceFiles', function() {
         it('should return sources file collection', function() {
-            $report = new Report([
+            $report = new CoverallsReport([
                 'sourceFiles' => new SourceFileCollection()
             ]);
             expect($report->getSourceFiles())->toBeAnInstanceOf('coverallskit\entity\collection\SourceFileCollection');
@@ -62,7 +62,7 @@ describe('Report', function() {
     describe('#validate', function() {
         context('when token empty', function() {
             beforeEach(function() {
-                $this->report = new Report([ 'token' => null ]);
+                $this->report = new CoverallsReport([ 'token' => null ]);
             });
             it('throw coverallskit\exception\RequiredException', function() {
                 expect(function() {
@@ -74,10 +74,10 @@ describe('Report', function() {
             beforeEach(function() {
                 $this->prophet = new Prophet();
 
-                $service = $this->prophet->prophesize('coverallskit\entity\ServiceInterface');
+                $service = $this->prophet->prophesize('coverallskit\entity\ServiceEntity');
                 $service->isEmpty()->willReturn(true);
 
-                $this->report = new Report([
+                $this->report = new CoverallsReport([
                     'token' => 'foo',
                     'service' => $service->reveal()
                 ]);
@@ -92,10 +92,10 @@ describe('Report', function() {
             beforeEach(function() {
                 $this->prophet = new Prophet();
 
-                $service = $this->prophet->prophesize('coverallskit\entity\ServiceInterface');
+                $service = $this->prophet->prophesize('coverallskit\entity\ServiceEntity');
                 $service->isEmpty()->willReturn(false);
 
-                $this->report = new Report([
+                $this->report = new CoverallsReport([
                     'token' => 'foo',
                     'service' => $service->reveal(),
                     'sourceFiles' => new SourceFileCollection()
@@ -118,16 +118,16 @@ describe('Report', function() {
                 new SourceFile(realpath(__DIR__ . '/../fixtures/bar.php'))
             ]);
 
-            $service = $this->prophet->prophesize('coverallskit\entity\ServiceInterface');
+            $service = $this->prophet->prophesize('coverallskit\entity\ServiceEntity');
             $service->isEmpty()->willReturn(false);
             $service->toArray()->willReturn([
                 'service_job_id' => '10',
                 'service_name' => 'travis-ci'
             ]);
 
-            $this->report = new Report([
+            $this->report = new CoverallsReport([
                 'token' => 'foo',
-                'repository' => new Repository(__DIR__ . '/../../'),
+                'repository' => new GitRepository(__DIR__ . '/../../'),
                 'service' => $service->reveal(),
                 'sourceFiles' => $sourceFiles
             ]);
@@ -157,16 +157,16 @@ describe('Report', function() {
                     new SourceFile(realpath(__DIR__ . '/../fixtures/bar.php'))
                 ]);
 
-                $service = $this->prophet->prophesize('coverallskit\entity\ServiceInterface');
+                $service = $this->prophet->prophesize('coverallskit\entity\ServiceEntity');
                 $service->isEmpty()->willReturn(false);
                 $service->toArray()->willReturn([
                     'service_job_id' => '10',
                     'service_name' => 'travis-ci'
                 ]);
 
-                $this->notSavedReport = new Report([
+                $this->notSavedReport = new CoverallsReport([
                     'token' => 'foo',
-                    'repository' => new Repository(__DIR__ . '/../../'),
+                    'repository' => new GitRepository(__DIR__ . '/../../'),
                     'service' => $service->reveal(),
                     'sourceFiles' => $sourceFiles
                 ]);
@@ -181,7 +181,7 @@ describe('Report', function() {
                 unlink($this->notSavedReport->getName());
             });
             it('should use the default name', function() {
-                expect($this->notSavedReport->getName())->toEqual(getcwd() . '/' . Report::DEFAULT_NAME);
+                expect($this->notSavedReport->getName())->toEqual(getcwd() . '/' . CoverallsReport::DEFAULT_NAME);
             });
             it('upload the report file', function() {
                 $this->prophet->checkPredictions();
@@ -196,16 +196,16 @@ describe('Report', function() {
                     new SourceFile(realpath(__DIR__ . '/../fixtures/bar.php'))
                 ]);
 
-                $service = $this->prophet->prophesize('coverallskit\entity\ServiceInterface');
+                $service = $this->prophet->prophesize('coverallskit\entity\ServiceEntity');
                 $service->isEmpty()->willReturn(false);
                 $service->toArray()->willReturn([
                     'service_job_id' => '10',
                     'service_name' => 'travis-ci'
                 ]);
 
-                $this->savedReport = new Report([
+                $this->savedReport = new CoverallsReport([
                     'token' => 'foo',
-                    'repository' => new Repository(__DIR__ . '/../../'),
+                    'repository' => new GitRepository(__DIR__ . '/../../'),
                     'service' => $service->reveal(),
                     'sourceFiles' => $sourceFiles
                 ]);