+++ src/configuration/Basic.php
@@ -12,8 +12,8 @@
 namespace coverallskit\configuration;
 
 use coverallskit\ReportBuilderInterface;
-use coverallskit\entity\Service;
-use coverallskit\entity\Repository;
+use coverallskit\entity\CIService;
+use coverallskit\entity\GitRepository;
 use coverallskit\Environment;
 use coverallskit\environment\AdaptorResolver;
 use Zend\Config\Config;
@@ -58,7 +58,7 @@ class Basic extends AbstractConfiguration
     }
 
     /**
-     * @return \coverallskit\entity\ServiceInterface
+     * @return \coverallskit\entity\ServiceEntity
      */
     public function getService()
     {
@@ -70,11 +70,11 @@ class Basic extends AbstractConfiguration
             $adaptor = $this->adaptorResolver->resolveByName($serviceName);
         }
 
-        return new Service($adaptor);
-     }
+        return new CIService($adaptor);
+    }
 
     /**
-     * @return \coverallskit\entity\Repository
+     * @return \coverallskit\entity\RepositoryEntity
      */
     public function getRepository()
     {
@@ -99,12 +99,12 @@ class Basic extends AbstractConfiguration
 
     /**
      * @param string $path
-     * @return \coverallskit\entity\Repository
+     * @return \coverallskit\entity\RepositoryEntity
      */
     private function repositoryFromPath($path)
     {
         $directory = $this->resolvePath($path);
-        $repository = new Repository($directory);
+        $repository = new GitRepository($directory);
 
         return $repository;
     }