+++ src/entity/CIService.php
@@ -17,10 +17,10 @@ use coverallskit\environment\AdaptorInterface;
 
 
 /**
- * Class Service
+ * Class CIService
  * @package coverallskit\entity
  */
-class Service implements ServiceInterface
+class CIService implements ServiceEntity
 {
 
     use AttributePopulatable;